Time: 0 ms (100.00%) | Memory: 7.9 MB (36.20%) - LeetSync

+++ 2050-count-good-numbers/count-good-numbers.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     const int m =1e9 + 7;
-    int findPower(long long a, long long b){
+    long long findPower(long long a, long long b){
         if(b == 0) return 1;
 
         long long  half  = findPower(a, b/2);
